only interface specs
